requirements.txt
alembic==1.6.5
appdirs==1.4.4
Babel==2.9.0
click==8.0.1
colorama==0.4.4
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
postgres==3.0.0
psycopg2-binary==2.9.1
psycopg2-pool==1.1
python-dateutil==2.6.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
virtualenv==20.4.7
Werkzeug==2.0.1
WTForms==2.3.3